[web] Fix search when an album only has symlinks
The uniquification would prevent albums that only contains photos that are
already present in another album from appearing in search results.
