+++ web/packages/new/photos/components/gallery/reducer.ts
@@ -33,7 +33,7 @@ import {
     getLatestVersionFiles,
     groupFilesByCollectionID,
 } from "../../services/file";
-import { sortFiles } from "../../services/files";
+import { sortFiles, uniqueFilesByID } from "../../services/files";
 import {
     isArchivedCollection,
     isArchivedFile,
@@ -762,29 +762,6 @@ const galleryReducer: React.Reducer<GalleryState, GalleryAction> = (
 
 export const useGalleryReducer = () =>
     useReducer(galleryReducer, initialGalleryState);
-
-/**
- * File IDs themselves are unique across all the files for the user (in fact,
- * they're unique across all the files in an Ente instance). However, we still
- * can have multiple entries for the same file ID in our local database because
- * the unit of account is not actually a file, but a "Collection File": a
- * collection and file pair.
- *
- * For example, if the same file is symlinked into two collections, then we will
- * have two "Collection File" entries for it, both with the same file ID, but
- * with different collection IDs.
- *
- * This function returns files such that only one of these entries (the newer
- * one in case of dupes) is returned.
- */
-export const uniqueFilesByID = (files: EnteFile[]) => {
-    const seen = new Set<number>();
-    return files.filter(({ id }) => {
-        if (seen.has(id)) return false;
-        seen.add(id);
-        return true;
-    });
-};
 
 /**
  * Compute archived collection IDs from their dependencies.
